requirements.txt
adrf==0.1.1
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
channels==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
Markdown==3.4.3
multidict==6.0.4
nextcord==2.4.2
pytz==2023.3
requests==2.31.0
spotipy==2.23.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==2.0.2
uvicorn==0.22.0
whitenoise==6.4.0
yarl==1.9.2